pyproject.toml
[project]
name = "fediviz"
version = "0.1.0"
description = ""
authors = [
    {name = "Ada", email = "github@adapdr.me"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "streamlit (>=1.41.1,<2.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "streamlit-extras (>=0.5.0,<0.6.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.0"
flake8 = "^7.1.1"